assignment 2 almost complete

+++ A2/src/GLSL.cpp
@@ -41,7 +41,7 @@ void checkVersion()
 	int major, minor;
 	major = minor = 0;
 	const char *verstr = (const char *)glGetString(GL_VERSION);
-	
+
 	if((verstr == NULL) || (sscanf(verstr, "%d.%d", &major, &minor) != 2)) {
 		printf("Invalid GL_VERSION format %d.%d\n", major, minor);
 	}
@@ -68,11 +68,11 @@ void printShaderInfoLog(GLuint shader)
 	GLint infologLength = 0;
 	GLint charsWritten  = 0;
 	GLchar *infoLog = 0;
-	
+
 	checkError(GET_FILE_LINE);
 	glGetShaderiv(shader, GL_INFO_LOG_LENGTH, &infologLength);
 	checkError(GET_FILE_LINE);
-	
+
 	if(infologLength > 0) {
 		infoLog = (GLchar *)malloc(infologLength);
 		if(infoLog == NULL) {
@@ -91,11 +91,11 @@ void printProgramInfoLog(GLuint program)
 	GLint infologLength = 0;
 	GLint charsWritten  = 0;
 	GLchar *infoLog = 0;
-	
+
 	checkError(GET_FILE_LINE);
 	glGetProgramiv(program, GL_INFO_LOG_LENGTH, &infologLength);
 	checkError(GET_FILE_LINE);
-	
+
 	if(infologLength > 0) {
 		infoLog = (GLchar *)malloc(infologLength);
 		if(infoLog == NULL) {
@@ -108,7 +108,7 @@ void printProgramInfoLog(GLuint program)
 		free(infoLog);
 	}
 }
-	
+
 char *textFileRead(const char *fn)
 {
 	FILE *fp;
@@ -130,7 +130,7 @@ char *textFileRead(const char *fn)
 			printf("error loading %s\n", fn);
 		}
 	}
-	
+
 	const char *t = content;
 	vector<int> v;
 	while(*t != '\0') {
@@ -163,7 +163,7 @@ int textFileWrite(const char *fn, const char *s)
 // https://www.tutorialspoint.com/utf-8-validation-in-cplusplus
 bool validUTF8(vector<int> &data) {
 	int cnt = 0;
-	for(int i = 0; i <data.size(); i++){
+	for(size_t i = 0; i <data.size(); i++){
 		int x = data[i];
 		if(!cnt) {
 			if((x >> 5) == 0b110) {
